+++ nessai/plot.py
@@ -49,6 +49,10 @@ def plot_live_points(live_points, filename=None, bounds=None, c=None,
 
     if c is not None:
         hue = df[c]
+        if np.all(hue == hue[0]):
+            logger.warning(
+                f'Selected hue variable: {c} is constant! Disabling.')
+            hue = None
     else:
         hue = None
 
@@ -302,6 +306,36 @@ def plot_trace(log_x, nested_samples, labels=None, filename=None):
     axes[-1].set_xlabel('log X')
     axes[-1].invert_xaxis()
 
+    if filename is not None:
+        fig.savefig(filename, bbox_inches='tight')
+        plt.close(fig)
+    else:
+        return fig
+
+
+def plot_histogram(samples, label=None, filename=None, **kwargs):
+    """Plot a histgram of samples.
+
+    Parameters
+    ----------
+    samples : array_like
+        Samples to plot.
+    label : str, optional
+        Label to the x axis.
+    filename : str, optional
+        Filename for saving the plot. If not specified, figure is returned.
+    kwargs :
+        Keyword arguments passed to `matplotlib.pyplot.hist`.
+    """
+    default_kwargs = dict(
+        histtype='step',
+        bins=auto_bins(samples)
+    )
+    default_kwargs.update(kwargs)
+    fig = plt.figure()
+    plt.hist(samples, **default_kwargs)
+    if label is not None:
+        plt.xlabel(label)
     if filename is not None:
         fig.savefig(filename, bbox_inches='tight')
         plt.close(fig)