+++ CHANGELOG.md
@@ -13,8 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added tests for `_NSIntegralState`.
 - Add code coverage upload
 - Added an example of using unbounded priors, `bilby_unbounded_priors.py`
-- Added `Rescale` reparameterisation that just rescales by a constant and does not require prior bounds. Also add tests for this reparameterisation.
+- Added `Rescale` reparameterisation that just rescales by a constant and does not require prior bounds. Also add
+tests for this reparameterisation.
 - Added more GW examples.
+- Added tests for `AugmentedFlowProposal`.
+- Added an example using `AugmentedFlowProposal`.
+- Added eggbox example.
 
 ### Changed
 
@@ -22,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `NestedSampler.plot_state` now has the keyword argument `filename` and the figure is only saved if it is specified.
 - Changed name from `_NSintegralState` to `_NSIntegralState`.
 - `nessai.model.Model` now inherits from `abc.ABC` and `log_prior` and `log_likelihood` are now `abstractmethods`. This prevents the class from being used without redefining those methods.
+- Updated `AumgentedFlowProposal` to work with current version of `FlowProposal`
+- Fix random seed unit tests.
+- Move `_NSIntegralState` and some functions from `posterior.py` to `evidence.py`
 
 ### Fixed
 
@@ -30,6 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Total sampling time is now correctly displayed when producing a state plot from a saved sampler.
 - Fixed a bug when using unbounded priors related to `Model.verify_model`
 - Fix inversion-split with `RescaleToBounds`
+- Fixed `AugmentedGWFlowProposal`.
+- Fixed a bug with `plot_live_points` when the hue parameter (`c`) was constant.
 
 
 ## [0.2.4] - 2021-03-08